requirements.txt
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==23.2
pytest==7.4.3
python-dateutil==2.8.2